moved ActivityFilter to a new file
