+++ oya-web/src/ActivityPage.tsx
@@ -1,12 +1,13 @@
-import { ChevronRight, ExpandMore, Search } from '@mui/icons-material';
+import { ChevronRight, ExpandMore } from '@mui/icons-material';
 import { TreeItem, TreeView } from '@mui/lab';
-import { Box, Button, InputAdornment, Paper, Stack, TextField, Typography } from '@mui/material';
+import { Box, Button, Paper, Stack, Typography } from '@mui/material';
 import React, { memo, useMemo, useState } from 'react';
 import { Activity } from './apiService';
 import ActivityEditor from './ActivityEditor';
 import useActivities from './useActivities';
 import { ActivityAdderDialog } from './ActivityAdder';
 import { DrawerHeader } from './Layout';
+import ActivityFilter, { ActivityFilterProps, ActivityFilters, filterActivities } from './ActivityFilter';
 
 export default function ActivityPage() {
   return (
@@ -23,6 +24,15 @@ const getAllTreeNodeIds = (activities: Activity[]) => {
   };
   activities.forEach(a => { getTreeIds(a); });
   return ids;
+};
+
+const initFilters: ActivityFilters = {
+  searchVal: '',
+  treeView: true,
+  hideSubActivities: true,
+  orderBychildrenLength: true,
+  order: 'name',
+  orderType: 'asc',
 };
 
 function ActivitiesTreeView() {
@@ -93,79 +103,6 @@ function ActivitiesTreeView() {
         onClose={() => setSelected('')}
       />
     </>
-  );
-}
-
-type ActivityFilters = {
-  searchVal: string,
-  treeView: boolean,
-  hideSubActivities: boolean,
-  orderBychildrenLength: boolean,
-  order: 'id' | 'name',
-  orderType: 'desc' | 'asc',
-};
-
-const initFilters: ActivityFilters = {
-  searchVal: '',
-  treeView: true,
-  hideSubActivities: true,
-  orderBychildrenLength: true,
-  order: 'name',
-  orderType: 'asc',
-};
-
-const filterActivities = (activities: Activity[], filters: ActivityFilters): Activity[] => {
-  const { searchVal, treeView, hideSubActivities, order, orderType, orderBychildrenLength } = filters;
-  const matches = (a: Activity) => a.name.trim().toLowerCase().includes(searchVal.toLowerCase().trim());
-  return [...activities].filter(a => {
-    if (treeView && hideSubActivities && a.parents.length > 0) return false;
-    if (!searchVal) return true;
-    if (treeView) {
-      if (
-        a.allChildren.some(matches) ||
-        (!hideSubActivities && a.allParents.some(matches))
-      ) return true;
-    } else {
-      if (a.allParents.some(matches)) return true;
-    }
-    return false;
-  }).sort((a, b) => {
-    if (orderBychildrenLength) {
-      const la = a.allChildIds.length;
-      const lb = b.allChildIds.length;
-      if (la !== lb) return lb - la;
-    }
-    let diff = 0;
-    if (a[order] < b[order]) diff = 1;
-    if (a[order] > b[order]) diff = -1;
-    return orderType === 'desc' ? diff : - diff;
-  });
-};
-
-type ActivityFilterProps = {
-  value: ActivityFilters;
-  onChange<T extends keyof ActivityFilters>(name: T, value: ActivityFilters[T]): void;
-};
-
-function ActivityFilter({ value, onChange }: ActivityFilterProps) {
-  return (
-    <Box>
-      <TextField
-        sx={{ mb: 2 }}
-        variant='outlined'
-        label='search'
-        fullWidth
-        value={value.searchVal}
-        onChange={(e) => onChange('searchVal', e.target.value)}
-        InputProps={{
-          endAdornment: (
-            <InputAdornment position='end'>
-              <Search />
-            </InputAdornment>
-          ),
-        }}
-      />
-    </Box>
   );
 }
